Support optional fields and Boolean-to-UInt conversion
Ensure compatibility with nullable structure fields by handling them properly in templates. Add conversion of Boolean values to UInt for consistency in function calls across platforms.

+++ wgpu4k-native/src/jvmMain/kotlin/Functions.jvm.kt
@@ -13,7 +13,7 @@ actual fun wgpuCreateInstance(descriptor: WGPUInstanceDescriptor?): WGPUInstance
 		?.let(::NativeAddress)?.let(::WGPUInstance)
 
 actual fun wgpuDevicePoll(device: WGPUDevice?, wait: Boolean, wrappedSubmissionIndex: WGPUWrappedSubmissionIndex?): Boolean
-	 = Functions.wgpuDevicePoll(device?.handler.adapt() ?: java.lang.foreign.MemorySegment.NULL, wait, wrappedSubmissionIndex?.handler.adapt() ?: java.lang.foreign.MemorySegment.NULL)
+	 = Functions.wgpuDevicePoll(device?.handler.adapt() ?: java.lang.foreign.MemorySegment.NULL, wait.toUInt(), wrappedSubmissionIndex?.handler.adapt() ?: java.lang.foreign.MemorySegment.NULL)
 		.toBoolean()
 
 actual fun wgpuSetLogCallback(callback: CallbackHolder<WGPULogCallback>?, userdata: NativeAddress?): Unit
